feat: split 2 reducers - for movies & modals
