+++ src/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from "redux"
+import moviesReducer from "./moviesReducer"
+import modalReducer from "./modalReducer"
+
+export default combineReducers({ 
+    movies: moviesReducer,
+    modal: modalReducer,
+})
