+++ src/reducers/modalReducer.js
@@ -1,0 +1,10 @@
+const initialState = {}
+
+const modalReducer = (state = initialState, action) => {
+    switch(action.type) {
+        default:
+            return state
+    }
+}
+
+export default modalReducer
